refactor: revert unuseful change
